Result and functions → NumberPrefix type
Rust has gained a few best practices since this crate was started (in 2014!) and it hasn’t kept up. One thing that’s now a bad idea is to use a type called Result for something that’s not an error type. And while NumberPrefix is _kind_ of an error type, it’s not enough of one to be called Result imho.

Changing the name made it more clear that the two floating decimal_prefix and binary_prefix functions should be moved onto the NumberPrefix type.